Support for damaging multiple tiles on attack
Renamed attack_pattern_single and attack_pattern_cardinal to generate_cardinal_pattern and generate_single_pattern as these functions are used for both attack and damage patterns

Added a direction parameter to generate_single_pattern which will make use of data in attack_pattern, in this case keeping the attack direction

Renamed calculate_attackable_tiles for same reasons as above, as well as some variable renames for clarity

Added calculate_damage_tiles which is called when the unit decides what tile it wants to attack and needs to know what tiles around it will be damaged.

Added last_damage_pattern var to Unit. Not particularly useful but keeps things consistent