Added first two location event data
Added an event for the starting area(Alzan), and for the first combat event(Alzan Forest). Events are functional and the combat event can be won.

Fixed bug in OverworldScreen where reaching a location was not setting the current event to the new location's event(went unnoticed because there was only one event so far)