Added Unit templating system
Worked out bugs in a system which will allow units to be spawned in from templates rather than having to manually enter each unit.

Fixed bug in Unit where ID generation was not actually generating unique IDs. The ID is now based on a combination of seconds and milleseconds as opposed to seconds

Tilemap will attempt to spawn reinforcements on init, which allows for reinforcements on round 0, which allows for templating

Tilemap.spawn_unit now returns the spawned unit

Units will now automatically write to unit data in the init function if they were spawned as a template unit